Former-commit-id: 6bef9d36bf012175b714e7947570506ee581708a

+++ sys-src/backend/app/connections/aws_s3.py
@@ -4,15 +4,11 @@ import os
 from io import BytesIO
 
 import numpy as np
+from app.exceptions import AWSError
 import cv2
 import boto3
 from botocore.exceptions import ClientError
 
-
-class AWSError(Exception):
-    def __init__(self, message, *args: object) -> None:
-        self.message = message
-        super().__init__(*args)
 
 # ====================================================================================================
 # Get AWS S3 connection
@@ -100,13 +96,14 @@ class S3Manager:
             _, buffer = cv2.imencode(
                 ".jpg", img
             )  # compresses opencv image and stores it in the memory buffer
-            img_as_bytes = BytesIO(buffer).getvalue()  # convert buffer to BytesIO object
+            img_as_bytes = BytesIO(
+                buffer
+            ).getvalue()  # convert buffer to BytesIO object
         except Exception:
             raise AWSError(message="Failed to encode image")
 
-
         # upload object to s3 bucket
-        try:        
+        try:
             connection = self._getConnection()
             bucket = self._getBucket(connection)
             obj = bucket.Object(objectKey)  # build object
@@ -118,7 +115,6 @@ class S3Manager:
         finally:
             self._closeConnection(connection)
 
-
     def deleteImageFromS3(self, objectKey):
         """
         Delete the s3 object with a specific key (name) from the s3 bucket
@@ -129,14 +125,16 @@ class S3Manager:
         try:
             connection = self._getConnection()
             bucket = self._getBucket(connection)
-            response = bucket.Object(objectKey).delete()
-            if response["ResponseMetadata"]["HTTPStatusCode"] != 204:
-                raise AWSError(message="Failed to delete Image from S3 bucket")
+            if objectKey == "defaultImage.jpg":
+                raise AWSError(message="Can not delete image with key defaultImage.jpg")
+            else:
+                response = bucket.Object(objectKey).delete()
+                if response["ResponseMetadata"]["HTTPStatusCode"] != 204:
+                    raise AWSError(message="Failed to delete Image from S3 bucket")
         except ClientError:
             raise AWSError(message="Failed to delete Image from S3 bucket")
         finally:
             self._closeConnection(connection)
-        
 
     def deleteAllImagesFromS3(self):
         """
@@ -146,6 +144,8 @@ class S3Manager:
             connection = self._getConnection()
             bucket = self._getBucket(connection)
             for obj in bucket.objects.all():
+                if obj.key == "defaultImage.jpg":
+                    continue
                 response = obj.delete()
                 if response["ResponseMetadata"]["HTTPStatusCode"] != 204:
                     raise AWSError(message="Failed to delete all Data from S3 bucket")
@@ -153,4 +153,3 @@ class S3Manager:
             raise AWSError(message="Failed to delete all Data from S3 bucket")
         finally:
             self._closeConnection(connection)
-
